chore: adding functions to test kots application install and upgrade.
implement two functions, one that installs a sample application on the
cluster and other that attempts to customize it after the cluster upgrade.

the idea is that after the install we will install the kots application
and customize it, once the cluster has been upgraded we check if the old
customization is still there and we attempt to customize once more the
app. if kots is not working properly after the cluster upgrade the new
customizations won't take effect.