bug: add --namespace flag to kubectl kots command
kubectl kots command, at least on v1.57.0, requires the namespace to be
specified in the command line (this is not true in the newest version).

due to its absense we have a few storage migration testgrids failing
with the following error:

attempting to customize kurl integration test application.
Error: failed to validate namespace: namespace is required
+ echo 'failed to customize kurl integration test application.'
+ exit 1